26-DoubleEntryPoint
on a first attempt I matched all of the calldata, but that could've been
circumvented pretty easily by just using another address to perform the
attack. On this one, I made it so transactions from the vault are
rejected, which is better in a sense (harder to circumvent), worse in
others (the CryptoVault now cannot move the funds of the legacy token,
least by sweeping), therefore still not ideal. But it's solved.